backend/requirements.txt
fastapi
uvicorn
pyswisseph
geopy
timezonefinder
google-genai
google-generativeai
firebase-admin
pydantic
razorpay
python-dotenv
resend # v1.0
